test: extract concurrenct embedded sub process test
